+++ src/PolylineEncoder.Net/Utility/Encoders/IPolylineEncoder.cs
@@ -6,9 +6,9 @@ namespace PolylineEncoder.Net.Utility.Encoders
 {
     public interface IPolylineEncoder
     {
-        string Encode(decimal latitude, decimal longitude);
+        string Encode(double latitude, double longitude);
         string Encode(IGeoCoordinate point);
-        string Encode(IEnumerable<Tuple<decimal, decimal>> points);
+        string Encode(IEnumerable<Tuple<double, double>> points);
         string Encode(IEnumerable<IGeoCoordinate> points);
     }
 }
